# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CentreCreate": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "maxLength": 250,
            "minLength": 1,
            "title": "Location",
            "type": "string"
          },
          "name": {
            "maxLength": 150,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "location"
        ],
        "title": "CentreCreate",
        "type": "object"
      },
      "CentreDetail": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "location",
          "is_active"
        ],
        "title": "CentreDetail",
        "type": "object"
      },
      "CentreUpdate": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "anyOf": [
              {
                "maxLength": 250,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 150,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "CentreUpdate",
        "type": "object"
      },
      "DiagnosticTestCreate": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 150,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "DiagnosticTestCreate",
        "type": "object"
      },
      "DiagnosticTestDetail": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "is_active"
        ],
        "title": "DiagnosticTestDetail",
        "type": "object"
      },
      "DiagnosticTestUpdate": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 150,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "DiagnosticTestUpdate",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "OfferingCreate": {
        "additionalProperties": false,
        "properties": {
          "price_minor": {
            "exclusiveMinimum": 0.0,
            "title": "Price Minor",
            "type": "integer"
          },
          "test_id": {
            "format": "uuid",
            "title": "Test Id",
            "type": "string"
          }
        },
        "required": [
          "test_id",
          "price_minor"
        ],
        "title": "OfferingCreate",
        "type": "object"
      },
      "OfferingDetail": {
        "additionalProperties": false,
        "properties": {
          "centre_id": {
            "format": "uuid",
            "title": "Centre Id",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "price_minor": {
            "title": "Price Minor",
            "type": "integer"
          },
          "test_id": {
            "format": "uuid",
            "title": "Test Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "centre_id",
          "test_id",
          "price_minor",
          "currency",
          "is_active"
        ],
        "title": "OfferingDetail",
        "type": "object"
      },
      "OfferingUpdate": {
        "additionalProperties": false,
        "properties": {
          "price_minor": {
            "exclusiveMinimum": 0.0,
            "title": "Price Minor",
            "type": "integer"
          }
        },
        "required": [
          "price_minor"
        ],
        "title": "OfferingUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "WebhookSecret": {
        "in": "header",
        "name": "X-Webhook-Secret",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Backend service for diagnostic test booking.",
    "title": "EVE Diagnostic Booking API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/centres/": {
      "get": {
        "operationId": "centres_centres__get",
        "parameters": [
          {
            "in": "query",
            "name": "location",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Centres Centres  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Centres",
        "tags": [
          "catalogue"
        ]
      },
      "post": {
        "operationId": "create_centre_centres__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CentreCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CentreDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create Centre",
        "tags": [
          "catalogue"
        ]
      }
    },
    "/centres/{centre_id}/": {
      "delete": {
        "operationId": "delete_centre_centres__centre_id___delete",
        "parameters": [
          {
            "in": "path",
            "name": "centre_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Centre Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete Centre",
        "tags": [
          "catalogue"
        ]
      },
      "get": {
        "operationId": "centre_centres__centre_id___get",
        "parameters": [
          {
            "in": "path",
            "name": "centre_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Centre Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CentreDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Centre",
        "tags": [
          "catalogue"
        ]
      },
      "patch": {
        "operationId": "patch_centre_centres__centre_id___patch",
        "parameters": [
          {
            "in": "path",
            "name": "centre_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Centre Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CentreUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CentreDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Patch Centre",
        "tags": [
          "catalogue"
        ]
      }
    },
    "/centres/{centre_id}/tests/": {
      "get": {
        "operationId": "offerings_centres__centre_id__tests__get",
        "parameters": [
          {
            "in": "path",
            "name": "centre_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Centre Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OfferingDetail"
                  },
                  "title": "Response Offerings Centres  Centre Id  Tests  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Offerings",
        "tags": [
          "catalogue"
        ]
      },
      "post": {
        "operationId": "create_offering_centres__centre_id__tests__post",
        "parameters": [
          {
            "in": "path",
            "name": "centre_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Centre Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferingDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create Offering",
        "tags": [
          "catalogue"
        ]
      }
    },
    "/centres/{centre_id}/tests/{test_id}/": {
      "delete": {
        "operationId": "delete_offering_centres__centre_id__tests__test_id___delete",
        "parameters": [
          {
            "in": "path",
            "name": "centre_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Centre Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "test_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Test Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete Offering",
        "tags": [
          "catalogue"
        ]
      },
      "patch": {
        "operationId": "patch_offering_centres__centre_id__tests__test_id___patch",
        "parameters": [
          {
            "in": "path",
            "name": "centre_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Centre Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "test_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Test Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferingUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferingDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Patch Offering",
        "tags": [
          "catalogue"
        ]
      }
    },
...
    "/tests/": {
      "get": {
        "operationId": "tests_tests__get",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Tests Tests  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tests",
        "tags": [
          "catalogue"
        ]
      },
      "post": {
        "operationId": "create_test_tests__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosticTestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticTestDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create Test",
        "tags": [
          "catalogue"
        ]
      }
    },
    "/tests/{test_id}/": {
      "delete": {
        "operationId": "delete_test_tests__test_id___delete",
        "parameters": [
          {
            "in": "path",
            "name": "test_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Test Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete Test",
        "tags": [
          "catalogue"
        ]
      },
      "patch": {
        "operationId": "patch_test_tests__test_id___patch",
        "parameters": [
          {
            "in": "path",
            "name": "test_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Test Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosticTestUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticTestDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Patch Test",
        "tags": [
          "catalogue"
        ]
      }
    }
  }
}